requirements.txt
fastapi
uvicorn
openai
requests
beautifulsoup4
PyPDF2
speechrecognition
edge-tts
python-dotenv
scikit-learn
langchain
langchain-openai
langchain-chroma
groq
gTTS
python-multipart
pydantic
pygame
pyttsx3
numpy
whisper
pyaudio
wave
